Add support for bidirectional text to KWIC
Added support for bidirectional text in the KWIC, such that phrases in Arabic or Hebrew are not shuffled, but the left and right side context appear correctly. The algorithm used is a gross simplification of a proper bidirectional algorithm, and doesn't work properly in a context where arabic words are enumerated with comma's in Latin script, punctuation appears at the end or start of the context, or Arabic numerals are used. However, for most cases, this is a big improvement.